Moved the modal background view controller into a separate file
